requirements.txt
pandas
python-dotenv==1.0.1
jsonpickle==3.0.3